perf: Optimize color changes
