+++ src/views/NFT/index.tsx
@@ -1,23 +1,11 @@
-import React, { useState } from 'react'
+import React from 'react'
 import styled from 'styled-components'
-import {
-  Appbar,
-  AppbarButton,
-  AppbarSticky,
-  HEADER_HEIGHT,
-} from '../../components/Appbar'
-import { ReactComponent as BackSvg } from '../../assets/svg/back.svg'
-import { ReactComponent as ShareSvg } from '../../assets/svg/share.svg'
-import { useHistoryBack } from '../../hooks/useHistoryBack'
+import { Appbar } from './components/appbar'
 import { Redirect, useParams, useRouteMatch } from 'react-router-dom'
 import { useNFTDetailApi } from './hooks/useNFTDetailApi'
 import { RoutePath } from '../../routes'
 import { Renderer } from './components/renderer'
 import { NftDetail } from './components/nftDetail'
-import { useObservable } from 'rxjs-hooks'
-import { fromEvent, tap } from 'rxjs'
-import { Box } from '@mibao-ui/components'
-
 const Container = styled.main`
   --max-width: 500px;
   position: relative;
@@ -27,24 +15,12 @@ const Container = styled.main`
 
 export const NFT: React.FC = () => {
   const { id } = useParams<{ id: string }>()
-  const goBack = useHistoryBack()
   const matchTokenClass = useRouteMatch(RoutePath.TokenClass)
   const { detail, failureCount, isLoading, refetch } = useNFTDetailApi(id, {
     isClass: matchTokenClass?.isExact,
   })
   const isNotFound =
     failureCount >= 3 || detail?.is_class_banned || detail?.is_issuer_banned
-  const [appbarBgOpacity, setAppbarBgOpacity] = useState(
-    Math.min(window.scrollY / 400, 1)
-  )
-
-  useObservable(() =>
-    fromEvent(window, 'scroll').pipe(
-      tap(() => {
-        setAppbarBgOpacity(Math.min(window.scrollY / 400, 1))
-      })
-    )
-  )
 
   if (isNotFound) {
     return <Redirect to={RoutePath.NotFound} />
@@ -52,31 +28,7 @@ export const NFT: React.FC = () => {
 
   return (
     <Container>
-      <AppbarSticky>
-        <Appbar
-          transparent
-          left={
-            <AppbarButton onClick={goBack}>
-              <BackSvg />
-            </AppbarButton>
-          }
-          right={
-            <AppbarButton transparent>
-              <ShareSvg />
-            </AppbarButton>
-          }
-        />
-        <Box
-          position="absolute"
-          top="0"
-          left="0"
-          bg="white"
-          h={`${HEADER_HEIGHT}px`}
-          w="full"
-          opacity={appbarBgOpacity}
-          zIndex={-1}
-        />
-      </AppbarSticky>
+      <Appbar />
 
       <Renderer detail={detail} />
       <NftDetail
